% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSensitivityTable.R
\name{strsplit1}
\alias{strsplit1}
\title{Generate a sensitivity table given a two input formula.}
\usage{
strsplit1(x, split)
}
\arguments{
\item{tableFormula}{A formula with two inputs that will define the sensitivity table values.}

\item{var1Base}{The base value of the first input of the formula.}

\item{var2Base}{The base value of the second input of the formula.}

\item{var1StepSize}{The step size of the first input of the formula. A negative value flips the order of values.}

\item{var2StepSize}{The step size of the second input of the formula. A negative value flips the order of values.}

\item{var1NumSteps}{The number of steps above and below the first input base.}

\item{var2NumSteps}{The number of steps above and below the second input base.}

\item{nDigits}{The number of digits the values will round to.}

\item{valueFormat}{What format the values should be. Options are "number" and "percent.}

\item{returnTable}{Whether a DT table should be returned or a matrix.}
}
\value{
A DT sensitivity table, or a matrix depending on returnTable.
}
\description{
Generate a sensitivity table given a two input formula.
}
\examples{
Generate a sensitivity table for monthly mortgage payments of a 30 year mortgage
based on the principle and interest rate.

pmt <- function(prinical, rate) {
  prinical * (rate/12)/(1-(1+rate/12)^-360)
}

GenerateSensitivityTable(pmt, 340000, .06, 10000, .0025, 3, 3, 2, 'number', FALSE)

}
